# backend/auth-service/src/main/java/com/company/platform/auth/claim/dto/response/ClaimResponse.java
package com.company.platform.auth.claim.dto.response;

import lombok.*;

@Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
public class ClaimResponse {
    private String id;
    private String claimCode;
    private String claimName;
}